+++ src/print.c
@@ -1,5 +1,6 @@
 #include "scop.h"
 #include "cvector.h"
+#include "s21_matrix.h"
 
 void t_vertex_print(t_vertex vertex)
 {
@@ -40,28 +41,28 @@ void t_face_transport_print(t_face_transport *t, char *label)
 void print_mesh(t_mesh *mesh)
 {
     printf("MESH OUTPUT:\n\n");
-    printf("vertices num: %d\n", cvector_size(mesh->vertices));
-    for (int i = 0; i < cvector_size(mesh->vertices); i++) {
+    printf("vertices num: %ld\n", cvector_size(mesh->vertices));
+    for (long unsigned int i = 0; i < cvector_size(mesh->vertices); i++) {
         t_vec3f_print(mesh->vertices[i], "vertex");
     }
 
-    printf("normals num: %d\n", cvector_size(mesh->normals));
-    for (int i = 0; i < cvector_size(mesh->normals); i++) {
+    printf("normals num: %ld\n", cvector_size(mesh->normals));
+    for (long unsigned int i = 0; i < cvector_size(mesh->normals); i++) {
         t_vec3f_print(mesh->normals[i], "normal");
     }
 
-    printf("textures num: %d\n", cvector_size(mesh->textures));
-    for (int i = 0; i < cvector_size(mesh->textures); i++) {
+    printf("textures num: %ld\n", cvector_size(mesh->textures));
+    for (long unsigned int i = 0; i < cvector_size(mesh->textures); i++) {
         t_vec2f_print(mesh->textures[i], "texture");
     }
 
-    printf("faces num: %d\n", cvector_size(mesh->faces));
-    for (int i = 0; i < cvector_size(mesh->faces); i++) {
+    printf("faces num: %ld\n", cvector_size(mesh->faces));
+    for (long unsigned int i = 0; i < cvector_size(mesh->faces); i++) {
         t_face_print(&mesh->faces[i], "face");
     }
 
-    printf("faces transport num: %d\n", cvector_size(mesh->faces_transport));
-    for (int i = 0; i < cvector_size(mesh->faces_transport); i++) {
+    printf("faces transport num: %ld\n", cvector_size(mesh->faces_transport));
+    for (long unsigned int i = 0; i < cvector_size(mesh->faces_transport); i++) {
         t_face_transport_print(&mesh->faces_transport[i], "face transport");
     }
 }
@@ -81,6 +82,7 @@ void t_pipeline_print(t_pipeline *p)
     o.r, o.l, o.b, o.t, o.n, o.f);
 
     t_camera_print(&p->m_camera);
+    /*
     s21_print_matrix(&p->m_WVPtransformation, "m_WVPtransformation");
     s21_print_matrix(&p->m_VPtransformation, "m_VPtransformation");
     s21_print_matrix(&p->m_WPtransformation, "m_WPtransformation");
@@ -88,4 +90,5 @@ void t_pipeline_print(t_pipeline *p)
     s21_print_matrix(&p->m_Wtransformation, "m_Wtransformation");
     s21_print_matrix(&p->m_Vtransformation, "m_Vtransformation");
     s21_print_matrix(&p->m_ProjTransformation, "m_ProjTransformation");
+    */
 }
